webapp/requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-configurations==2.5.1 ; python_version >= "3.12" and python_version < "4.0"
django-cors-headers==4.7.0 ; python_version >= "3.12" and python_version < "4.0"
django-filter==25.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.2.1 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework-simplejwt==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.16.0 ; python_version >= "3.12" and python_version < "4.0"
drf-yasg==1.21.10 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
joblib==1.5.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.3.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
pillow==11.2.1 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.2.9 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "pypy"
psycopg==3.2.9 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
scikit-learn==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
scipy==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
threadpoolctl==3.6.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version == "3.12"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
